incorporate joel's feedback
added a new map problem. Adapted the old map problem into a foreach problem because that was more appropriate.
Problem6 still needs a solution to be written.
all function names are now in lowercase
problem6 is now meant to have the map be built dynamically
problem5 now assumes that problem2 will be called before it as a precondition